Update unit tests to use saucelabs.
