+++ package.json
@@ -107,6 +107,7 @@
     "karma-jasmine-matchers": "^4.0.1",
     "karma-rollup-preprocessor": "^7.0.5",
     "karma-safari-launcher": "^1.0.0",
+    "karma-sauce-launcher": "^4.3.5",
     "karma-spec-reporter": "0.0.32",
     "lint-staged": "^9.4.2",
     "prettier": "^1.16.4",
